requirements.txt
japronto==0.1.2
redis==3.0.1
PyYAML==3.13
